Moving checkers also to the warnings_checker file
